Check if test database exists before deleting it
Otherwise we will panic if we attempt to delete the non-existent file.
We could also just remove the .expect(), but then we wouldn't know if
there were other issues like permissions.

+++ src/tests_main.rs
@@ -7,7 +7,9 @@ use std::fs;
 
 //set up rocket & empty test database
 fn setup_test_rocket() -> rocket::Rocket {
-    fs::remove_file("test_database.db").expect("bug: cannot delete old database");
+    if Path::new("test_database.db").exists() {
+        fs::remove_file("test_database.db").expect("bug: cannot delete old database");
+    }
     init_database_file("test_database.db");
 
     let mut database_config = HashMap::new();
